+++ src/index.ts
@@ -81,7 +81,7 @@ const optionsConst = `
         --externalUrl                     Pass in external url to check valid transaction. This parameter is mandatory for mainnet (if testnet is false).  Daemon mode only.
         --externalUrlMethod               Determine what http method to use when calling the url passed in the --externalUrl option. Accepts GET or POST. Defaults to GET method. Daemon mode only. 
         --runOnce                         Run the daemon command one time. Check for a new transactions to sign once, and then exit the process. Daemon mode only.
-        --wallet                          If runOnce enabled, fetch and sign only for this wallet id.
+        --wallets                         If runOnce enabled, fetch and sign only for these wallet ids.
         --transactionIds                  If runOnce enabled, sign only transactions from defined comma-separated list.
 `
 
@@ -143,7 +143,7 @@ const startup = async () => {
         type: 'boolean',
         default: false,
       },
-      wallet: {
+      wallets: {
         type: 'string'
       },
       transactionIds: {
@@ -176,7 +176,7 @@ const startup = async () => {
         flags.externalUrlMethod as ExternalUrlMethod,
         flags.period as number,
         flags.runOnce as boolean,
-        flags.wallet as string,
+        (flags.wallets as string)?.split(','),
         (flags.transactionIds as string)?.split(','),
       )
       break